StudentModalityController moved to new folder Student 	renamed:    app/Http/Controllers/api/StudentModalityController, -> app/Http/Controllers/api/Student/StudentModalityController.php 	modified:   routes/api/v1.php 	modified:   tests/Feature/Student/StudentModalityControllerTest.php
